+++ src/axis/leg/leg_coordinates/leg_coordinates.rs
@@ -7,11 +7,11 @@ use crate::{
 pub trait LegCoordinates: LegQuantities {
     fn take_region(limit_price: TickPosV2, price: TickPosV2) -> TakeRegion;
 
-    fn start<K, const BITS: usize>() -> DerivedPosition<K, BITS>
+    fn start<K, const BITS: u16>() -> DerivedPosition<K, BITS>
     where
         K: InnerVal;
 
-    fn end<K, const BITS: usize>() -> DerivedPosition<K, BITS>
+    fn end<K, const BITS: u16>() -> DerivedPosition<K, BITS>
     where
         K: InnerVal;
 }
